Move RSVP list button to view page
