+++ app/editor/[entity]/event/actions.ts
@@ -1,12 +1,11 @@
 'use server';
 
 import { redirect } from 'next/navigation';
-import { ActionResponse, Event, EventRsvp } from '@/definitions/types';
+import { ActionResponse, Event } from '@/definitions/types';
 import { DBEventHelper } from '@/dynamodb/DBEventHelper';
-import { DBEventRsvpHelper } from '@/dynamodb/DBEventRsvpHelper';
 import { parseWithSchema } from '@/editor/schema/helpers';
 import { SettingsFormData } from '@/editor/types';
-import { toActionFailure, toActionSuccess } from '@/helpers/action';
+import { toActionFailure } from '@/helpers/action';
 import { eventSchema } from './schema';
 
 export type EventFormData = SettingsFormData &
@@ -51,13 +50,4 @@ export async function publishEvent(
   }
 
   redirect(`/${shortId}`);
-}
-
-export async function fetchRsvpList(
-  shortId: string
-): Promise<ActionResponse<EventRsvp[]>> {
-  const helper = new DBEventRsvpHelper();
-
-  const items = await helper.getItems({ shortId });
-  return toActionSuccess(items);
 }
